+++ frontend/src/App.js
@@ -1,25 +1,12 @@
 import React from 'react';
 import logo from './logo.svg';
 import './App.css';
-
+import NavBar from './components/navbar'
 
 function App() {
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <h1>
-          INDEX PAGE 
-        </h1>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+      <NavBar />
     </div>
   );
 }
